+++ spec/fixtures/cookbooks/oneview_test_api300_synergy/recipes/switch_patch.rb
@@ -1,4 +1,8 @@
-# (c) Copyright 2016 Hewlett Packard Enterprise Development LP
+#
+# Cookbook Name:: oneview_test_api300_synergy
+# Recipe:: switch_patch
+#
+# (c) Copyright 2017 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -8,17 +12,14 @@
 # under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
+#
 
-OneviewCookbook::ResourceBaseProperties.load(self)
-
-default_action :update
-
-property :associated_ethernet_network, String
-
-action :update do
-  OneviewCookbook::Helper.do_resource_action(self, :ConnectionTemplate, :create_or_update)
-end
-
-action :reset do
-  OneviewCookbook::Helper.do_resource_action(self, :ConnectionTemplate, :reset)
+oneview_switch 'Switch1' do
+  client node['oneview_test']['client']
+  api_version 300
+  api_variant 'C7000'
+  operation 'test'
+  path 'test/'
+  value 'TestMessage'
+  action :patch
 end